pyproject.toml
[tool]
[tool.poetry]
name = "duniterpy"
version = "0.60.0"
description = "Python library for developers of Duniter clients"
authors = ["inso <insomniak.fr@gmail.com>", "canercandan", "s_b"]
maintainers = ["vit <vit@free.fr>", "Moul <moul@moul.re>"]
readme = "README.md"
license = "GPL-3.0-or-later"
homepage =  ""
repository = "https://git.duniter.org/clients/python/duniterpy"
documentation = "https://clients.duniter.io/python/duniterpy/index.html"
keywords = ["g1", "duniter", "cryptocurrency", "librecurrency", "library"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
    "Intended Audience :: Developers",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.6.8"
aiohttp = "^3.6.3"
jsonschema = "^3.2.0"
pypeg2 = "^2.15.2"
attrs = "^20.2.0"
base58 = "^2.0.0"
libnacl = "^1.7.2"
pyaes = "^1.6.1"
graphql-core = "^3.1.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.790"
pylint = "^2.6.0"
sphinx = "^3.2.1"
sphinx_rtd_theme = "^0.5.0"
